Fixed weirdo vi header include error
A while back, my local vi installation had a habit
of swapping lines if I hit the enter key right after
opening a file. This is probably related to a different
error in my vi configuration, which seems otherwise benign.

The upshot of this line swapping is that I have gotten a
host of absolutely bizzaro errors which make no sense. In
this case, the header leading #ifndef and #define lines
were swapped.

Very disconcerting when something which works appears to
break for no reason.